% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_CleanPathfinder.R
\name{dge_CleanPathfinder}
\alias{dge_CleanPathfinder}
\title{Clean up the output of dge_RunPathfinder to make nicer bubble plots}
\usage{
dge_CleanPathfinder(pathRes, wrap = 35, foldChange = 1, geneCount = 20)
}
\arguments{
\item{pathRes}{Dataframe, output of dge_RunPathfinder().}

\item{wrap}{Numeric, max number of characters in a pathway name before
wrapping to a new line.}

\item{foldChange}{Numeric, keep only pathways with a fold enrichment \strong{over}
this number.}

\item{geneCount}{Numeric, keep only pathways with a gene count \strong{over} this
number.}
}
\value{
Dataframe with the above cleaning steps applied.
}
\description{
Small cleaning function to do things like wrap long pathway names, and filter
on fold enrichment scores and gene counts to get rid of irrelevant pathways.
}
