% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_CollapseKeggLists.R
\name{dge_CollapseKeggLists}
\alias{dge_CollapseKeggLists}
\title{Collapse a set of KeggGeneLists into one list of vectors}
\usage{
dge_CollapseKeggLists(keggResList)
}
\arguments{
\item{keggResList}{List of results from running multiple
dge_GetKeggGeneLists() commands, likely from vectorizing it over multiple
dge_RunPathview requests.}
}
\value{
A single list of named vectors, like the output of
dge_GetKeggGeneLists: one vector per pathway, contains all sig genes in the
pathway.
}
\description{
The idea here is that if you run a DGE experiment with multiple comparisons,
you'll probably be vectorizing your whole analysis. However, at the end, when
you want to show heatmaps of significant genes in pathways, it'll be helpful
to take your list of significant KEGG terms (one set per comparison) and make
a single, main list that contains all significant genes for all signficiant
pathways, so that your heatmap (which will contain all samples) is
representative. This function does exactly that.
}
