% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_GetKeggGeneLists.R
\name{dge_GetKeggGeneLists}
\alias{dge_GetKeggGeneLists}
\title{Get a list of all significantly altered genes in all significant pathways}
\usage{
dge_GetKeggGeneLists(pathRes)
}
\arguments{
\item{pathRes}{Dataframe, the output of dge_RunPathfinder().}
}
\value{
A named list of character vectors containing gene names, with each
vector named by its KEGG pathway.
}
\description{
Returns one list per pathway, from the output of dge_RunPathfinder().
}
