% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_GetMatrix.R
\name{dge_GetMatrix}
\alias{dge_GetMatrix}
\title{Get CPM-corrected count matrix out of a DGEList, then filter on genes of interest}
\usage{
dge_GetMatrix(DGEList, genes = NULL, scale = "none", colnames = NULL)
}
\arguments{
\item{DGEList}{A DGEList object, usually returned when you run
dge_OneFactor() with retainDGEList = TRUE.}

\item{genes}{Character vector of gene names, in the same format as the
rownames from your DGEList. If you're unsure what these will look like in
your own data, run this function keeping genes = NULL and then peek
(head()) at the output.}

\item{scale}{One of "none", "cols", or "rows"; should centering and scaling
be applied to the data, and if so, should it be row-wise or column-wise?}

\item{colnames}{Optional vector of new colnames. This will be directly
replaced 1:1, so it needs to be the same length as the number of columns in
the output, and in the same order. Useful for when your sample names are
gross and you want to replace them with something nicer for the resulting
heatmap.}
}
\value{
A matrix with (selected) genes as rows, samples as columns, and
values as CPM-adjusted counts with optional scaling.
}
\description{
Get the DGEList out of a dge_OneFactor result, and return a matrix of
(selected) genes.
}
\examples{

  dgeRes <- system.file("extdata", "dgeRes_full.Rds",
                      package = "seqHelpers") |>
    readRDS()

  dge_GetMatrix(dgeRes$DGEList)

}
