% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_OneFactor.R
\name{dge_OneFactor}
\alias{dge_OneFactor}
\title{Perform Differential Gene Expression analysis on a one-factor experiment}
\usage{
dge_OneFactor(
  HitCountsMatrix,
  Groups,
  refGroup = NULL,
  convertNames = FALSE,
  convertFrom = NULL,
  convertTo = NULL
)
}
\arguments{
\item{HitCountsMatrix}{A matrix of hit counts, where each row is a gene and
each column is a sample. Note that this should be de-duplicated before
running through this function; i.e. each gene should be UNIQUE, and if they
aren't you should add the non-unique reads together with dplyr before you
start in on this!}

\item{Groups}{A factor showing what group each sample belongs to. The length
of this factor must match the number of columns in the HitCountsMatrix
exactly}

\item{refGroup}{The control group that each other group in the Groups factor
will be compared against. If not specified (refGroup = NULL), defaults to
the first level in the factor.}

\item{convertNames}{Boolean specifying whether names should be converted.
Note that this requires an internet connection, so keep false if you are
offline. Typically, this is used to convert the stable Ensembl ID that
GeneWiz gives us into a more useful MGI symbol for downstream analysis.}

\item{convertFrom}{a string specifying a biomart object to convert the names
from. Popular choices are "entrezgene_id" and "ensembl_gene_id".}

\item{convertTo}{a string specifying a biomart object (or character vector
for several) to convert names to. Popular choices are "mgi_symbol".}
}
\description{
This function handles a "default" call of edgeR to perform differential gene
expression testing on a count matrix file. The experiment must be a
one-factor design (this function isn't set up to handle multi-factor problems
yet), but can have any number of levels in that design, as long as they are
compared to a single reference group. Note that edgeR itself can handle much
more complex designs through limma, but having the ability to catch that
complexity in this function call is not so straightforward, and so very
complex experiments should probably be run more manually than this.
}
