% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_PlotHeatmap.R
\name{dge_PlotHeatmap}
\alias{dge_PlotHeatmap}
\title{Plot CPM heatmap with sensible defaults}
\usage{
dge_PlotHeatmap(
  mat,
  scale = "row",
  cluster_rows = TRUE,
  show_cluster_rows = FALSE,
  ...
)
}
\arguments{
\item{mat}{A matrix, likely the output of dge_GetMatrix() (not neccesarily,
though).}

\item{scale}{Arg for ggheatmap, defaulting to "row" (gene-wise)}

\item{cluster_rows}{Arg for ggheatmap, defaulting to TRUE (yes, cluster genes)}

\item{show_cluster_rows}{Arg for ggheatmap, defaulting to FALSE (no, don't print the dendrogram)}

\item{...}{Additional args to pass to ggheatmap}
}
\value{
A ggheatmap object
}
\description{
Useful for getting a quick heatmap out of the output of dge_GetMatrix
}
