% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge_PrepKeggData.R
\name{dge_PrepKeggData}
\alias{dge_PrepKeggData}
\title{Convert DEG results tables to something Pathview can work with}
\usage{
dge_PrepKeggData(
  dgeRes,
  geneNames = .data$Newid,
  pval = 0.05,
  enforceSentenceCase = FALSE,
  deduplicate = FALSE
)
}
\arguments{
\item{dgeRes}{Results table as returned by dge_OneFactor().}

\item{geneNames}{Tidy Eval selection of the column in dge_OneFactor that
contains the gene names. Usually Newid if you converted names to MGI
symbols and GeneID if you did not.}

\item{pval}{Numeric giving the P Value cutoff to use. Will keep only rows
with adjusted P value \emph{less than} this.}

\item{enforceSentenceCase}{Boolean; should gene names be forced into sentence
case (capital first letter only)?}

\item{deduplicate}{Boolean; should genes be de-duplicated by keeping only the
most significant instance of each one?}
}
\value{
A dataframe suitable for plotting KEGG pathways with plotKEGG().
}
\description{
Helper function to take the output table from running dge_OneFactor(), filter
on an adjusted P value cutoff, and format it to something pathview can use.
Also can optionally enforce sentence case for mouse gene IDs (useful when you
get data from an informatics core that defaults to human-style gene names)
and deduplicating by selecting the most significant fold change when
duplicates are present.
}
