% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKEGG.R
\name{plotKEGG}
\alias{plotKEGG}
\title{Create a KEGG pathway diagram with DEG data overlay.}
\usage{
plotKEGG(
  keggData,
  pathway,
  species = "Mouse",
  inputDir = "KEGG Inputs/",
  plotName = NULL,
  geneCols = list(low = "cyan", mid = "white", high = "firebrick"),
  cpdCols = list(low = "blue", mid = "gray", high = "yellow")
)
}
\arguments{
\item{keggData}{A dataframe containing MGI symbol-formatted gene IDs,}

\item{pathway}{A string giving the pathway ID for the pathway of interest, in
the format "00001". No species code is required.}

\item{species}{A string specifying the species to use. Currently, only
"Mouse" is supported.}

\item{inputDir}{A path to a directory where pathview can download the KEGG
files. Directory must exist before running this!}

\item{plotName}{An optional string to append to the output image file. Useful
when vectorizing so each subsequent output doesn't overwrite the previous
one.}

\item{geneCols}{A 3-length list giving the low, medium, and high colour
values to map to gene expression. Please keep them in that order, names are
optional.}

\item{cpdCols}{As geneCols, but for compound values}
}
\value{
Nothing - instead, writes the output png to the working directory.
Would love to have it not do that but this seems to be a feature of
pathview.
}
\description{
Create a KEGG pathway diagram with DEG data overlay.
}
